Add a workspace endpoint for schema-level metadata
`oxen schemas add <path> -m '{...}'` can set metadata on a data frame's
schema, but only the per-column half of that had an HTTP equivalent: the
/schemas service is GET-only, and the data frame PUT reads nothing but
is_indexed. So a client could set a column's render func but had nowhere to
persist file-level schema settings.

Adds PUT /workspaces/{id}/data_frames/schema/{path}, taking `{"metadata": ...}`
and staging it into the workspace like the column endpoint does — no commit
required, which matters for data frames that live in a workspace and are never
committed.

Staging such an edit is identical whether it targets a column or the schema:
resolve the file node from the staged db or the commit tree, stage its parent
directories, reconcile the schema against the workspace's DuckDB table, rehash
and upsert. Only the mutation differs, so that is extracted into
stage_schema_metadata_update and both callers pass a closure — otherwise the
two copies would drift, and this logic is subtle enough that drift would be
quiet.

Test covers the case the two levels make easy to get wrong: writing schema
metadata leaves per-column metadata (render funcs) intact.
